pyproject.toml
[tool.poetry]
name = "image-processing-pipeline-poc"
version = "1.0.0"
description = ""
authors = ["Mykhailo Chalyi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12.4"
pypdf = "^4.3.1"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
langgraph = "^0.2.4"
langchain-community = "^0.2.12"
rapidocr-onnxruntime = "^1.3.24"
pymupdf = "^1.24.9"
boto3 = "^1.35.0"
tiktoken = "^0.7.0"
amazon-textract-caller = "^0.2.4"
amazon-textract-textractor = "^1.8.2"
pdf2image = "^1.17.0"
rich = "^13.7.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
